Neaten the history listing code
I'd like to reduce the amount of SQL in the database code. Make it as
generic as possible, and later on perhaps expose a generic "execute"
function.

This function can be used by analysis commands, and the SQL can live
there - rather than database.rs being a huge bag of SQL.