Gate the various functionality using features
Instead of forcing the user(s) to build a binary that contains both the
client and the server, we can use features to only compile what is
needed.

Furthermore, not all users want the "syncing" functionality in their
client and this allows the user to disable that functionlity from even
being compiled and gives more piece of mind it will never attempt to
make external calls.

The dependency graph shrinks dramatically as well when only building the
client without "sync".

For now, the defaults still include all of the features to retain
existing functionality by default.