fix: resolve some issues with install.sh (#188)
* feat: add command checks to install script

Some people don't use sudo or have curl, sed or wget installed by default.
This adds a check, if sudo, curl and sed are installed and changes wget
to curl as this is the mostly used command in the script. If sudo is not
installed it uses su.

Closes: #175

* fix: add newline to printf commands

The printf was missing a newline at the end, which resulted in the eval
line being appended directly to the end of the previous line.