Replace create or replace trigger syntax
PostgreSQL 13 doesn't support the 'create or replace' trigger syntax
yet. As PostgreSQL 13 was released fairly recently, in September 2020,
is still supported and will be until 2025, it might make sense to use
SQL statements that are supported by PostgreSQL 13.

I have rewritten the SQL in the migration to drop the trigger and
re-create it, wrapped in a transaction, instead of using the 'create or
replace' syntax. This works for me, but I obviously only have a single
atuin server instance running.