Make `atuin account delete` void session + key
When I deleted my Atuin.sh account, I found the command did not delete
the key or session files.

This caused `bail!`s from Atuin when querying the session status, or
synchronising.

Of course, one shouldn't expect Atuin to be able to synchronise or query
with the server when the account is deleted, but the relevant files were
still present, so the behaviour was different to what we'd expect.

I discussed this with @ellie on Discord, and we came to the conclusion
that I should open a PR, and submit a patch.

I'm not sure how well this PR fits in with the borrow checker, but I've
run tests, a Nix build, and a Cargo build - all goes well. I have not
tested on macOS or Windows; only NixOS.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>