fix(bash): prefix "__atuin_" to avoid variable conflicts
Because the function "__atuin_history" executes an arbitary user
command for "enter_accept", the local variable names should be
carefully chosen.  A local variable can shadow a global variable that
the user wants to use when there is a name conflict.  To avoid such a
situation we try to namespace the variables used by atuin by prefixing
"__atuin_".