fix(bash,zsh): fix quirks on search cancel (#1483)
* fix(bash): preserve the line content on search cancel

In the current implementation for Bash, the line content is lost when
the user cancels the atuin search by pressing ESC, C-g, or Down at the
bottom line.  This is because the line content is set to the empty
string returned by atuin on the cancellation of the search.

In the integrations for other shells, zsh and fish, the empty output
is properly handled so that the line content is preserved.  This patch
makes the behavior in Bash consistent with that in zsh and fish, i.e.,
we do nothing when the atuin search returns an empty output.

* fix(zsh): ignore confusing line `__atuin_accept__:*` on search cancel