feat: enable enhanced keyboard mode
This enabled the Kitty Keyboard Protocol

Read more here: https://sw.kovidgoyal.net/kitty/keyboard-protocol/

No change on unsupported terminals, but means in the future we can be
more creative with keybinding depending on terminal.

Tested on Alacritty and events come through with all modifiers
supported.

Will be useful for #193