fix(bash): work around custom IFS
When the user sets a custom value of IFS, the up keybinding may fail
because the option `--shell-up-key-binding` passed to
`__atuin_history` is broken by the word splitting of the shell.  For
example, when the user sets IFS=-, `atuin` called from __atuin_history
receives `shell up key binding <content>` as the search string.

$ IFS=-
$ echo [up] # <-- this does not work as expected

Note that the problem only happens with the plain Bash without ble.sh.
The problem does not arise within ble.sh because ble.sh separates the
user environment and the line-editor environment by saving/restoring
the shell settings.  The keybindings are processed in the line-editor
environment, so the custom IFS set by the user does not affect it.

In this patch, we properly quote the arguments to the atuin command.