fix(bash): add guard for double initialization
In bash, it is customary to reload the settings by sourcing `.bashrc`
again after modifying it.  In such a case, `eval "$(atuin init bash)"`
is executed again.  This registers duplicate hooks to
`preexec_functions` and `precmd_functions`.  To prevent this in this
patch, we introduce an include guard, so that the initialization is
not performed more than once.