refactor(search): check ctrl for j and k bindings
In this patch, we apply the rules for vim-modes only on `j` and `k`
without the Control modifier.

With the current set of keybindings, the current implementation does
not cause a problem because `j` and `k` behave the same as `C-j` and
`C-j` within the vim-normal mode.  However, this is just accidental
and can be broken by future changes.  To make these keybindings
orthogonal, it is better to explicitly check the Control modifier.