fix(bash/preexec): erase the prompt last line before Bash renders it
Fixes https://github.com/atuinsh/atuin/issues/1668

When the prompt becomes longer after "enter_accept", Bash still uses
the previous shorter prompt, so the extra characters in the new prompt
drawn by Atuin is left in the terminal display.  In this patch, we
remove the last line of the prompt drawn by Atuin so that it doesn't
interfere with the last line of the prompt drawn by Bash.